requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
httpx==0.25.2
PyPDF2==3.0.1
python-multipart==0.0.6

# Python version specification
python-version>=3.8
